requirement.txt
streamlit
pandas
numpy
matplotlib
seaborn
plotly
folium
streamlit-folium
scikit-learn
joblib
geopy.geocoders
plotly.express
sklearn